review PR - remove dead codes - revert outline removal for asseccibility - unify common data - sort import
