--- src/components/shared/TitleCardSection.tsx
@@ -1,1 +1,0 @@
-export {}
